sql: move ConvertBatchError to sqlbase
This is to allow it to be eventually called from distsql.
