+++ pkg/sql/sqlbase/errors.go
@@ -19,11 +19,14 @@
 package sqlbase
 
 import (
+	"fmt"
 	"strings"
 
+	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/pkg/errors"
 )
 
@@ -204,4 +207,59 @@ func errHasCode(err error, code string) bool {
 		return c == code
 	}
 	return false
+}
+
+// ConvertBatchError returns a user friendly constraint violation error.
+func ConvertBatchError(tableDesc *TableDescriptor, b *client.Batch) error {
+	origPErr := b.MustPErr()
+	if origPErr.Index == nil {
+		return origPErr.GoError()
+	}
+	index := origPErr.Index.Index
+	if index >= int32(len(b.Results)) {
+		panic(fmt.Sprintf("index %d outside of results: %+v", index, b.Results))
+	}
+	result := b.Results[index]
+	var alloc DatumAlloc
+	if _, ok := origPErr.GetDetail().(*roachpb.ConditionFailedError); ok {
+		for _, row := range result.Rows {
+			// TODO(dan): There's too much internal knowledge of the sql table
+			// encoding here (and this callsite is the only reason
+			// DecodeIndexKeyPrefix is exported). Refactor this bit out.
+			indexID, key, err := DecodeIndexKeyPrefix(&alloc, tableDesc, row.Key)
+			if err != nil {
+				return err
+			}
+			index, err := tableDesc.FindIndexByID(indexID)
+			if err != nil {
+				return err
+			}
+			vals, err := MakeEncodedKeyVals(tableDesc, index.ColumnIDs)
+			if err != nil {
+				return err
+			}
+			dirs := make([]encoding.Direction, 0, len(index.ColumnIDs))
+			for _, dir := range index.ColumnDirections {
+				convertedDir, err := dir.ToEncodingDirection()
+				if err != nil {
+					return err
+				}
+				dirs = append(dirs, convertedDir)
+			}
+			if _, err := DecodeKeyVals(&alloc, vals, dirs, key); err != nil {
+				return err
+			}
+			decodedVals := make([]parser.Datum, len(vals))
+			var da DatumAlloc
+			for i, val := range vals {
+				err := val.EnsureDecoded(&da)
+				if err != nil {
+					return err
+				}
+				decodedVals[i] = val.Datum
+			}
+			return NewUniquenessConstraintViolationError(index, decodedVals)
+		}
+	}
+	return origPErr.GoError()
 }
